requirements.txt
numpy
pandas
scikit-learn
matplotlib
python-dotenv
google-generativeai
